Merge in WLMDB/wlmdb from Demo-bug-fixes to master

* commit '643f3f25bb7e844de29653040fb6d08bb64bffa8':
  DBS-3801 unmanage bug fix

+++ server/src/operations/discover-operations.ts
@@ -2088,7 +2088,7 @@ async function unmanageDatabaseInstance(
         // When all database instances are removed, the EC2 ceases to be a
         // managed resource, since  we aren't managing any SQL Server instance.
         // So we need to remove the EC2 resource from wlmdb.resource table.
-        if (postDeleteDatabaseInstances.length <= 0) {
+        if (postDeleteDatabaseInstances.length <= 0 && !isDemoFlow) {
             deleteResource(accountId, resourceId, credentialsId);
         }
     }
